refactor: split command sending into SitemapPageViewModel+Commands.swift
Extracts the MARK: Command Sending section and command-state helpers
into a dedicated extension file to bring SitemapPageViewModel.swift
under the 1000 non-blank/comment line SwiftLint limit.

Removes private from the nine properties that the new file needs
(commandDispatcher, openAPIService, defaultSitemap, queuedCommands,
commandStates, commandStateVersions, commandStateResetTasks,
sliderValueOverrides, sliderOverrideResetTasks); all remain internal
to the app target.
